feat: change vote_receipt to Vec<u8> for correct env::verify journal format
env::verify internally casts journal to &[u8] via bytemuck. Using Vec<u8>
matches the raw RISC0 journal bytes from the pre_tx hook, avoiding any
u32-to-u8 truncation issues. Proven working in test-identity E2E flow.

Changes:
- multisig_program: Vec<u32> → Vec<u8> for vote_receipt in propose/approve/reject
- multisig_core: matching type changes
- multisig_client: matching type changes

+++ multisig_core/src/lib.rs
@@ -62,8 +62,8 @@ pub enum Instruction {
         proposal_index: u64,
         /// Member NPKs (needed by pre_tx hook for ZK proof)
         members: Vec<[u8; 32]>,
-        /// RISC0 receipt from vote_circuit proving proposer membership
-        vote_receipt: Vec<u32>,
+        /// RISC0 receipt from vote_circuit proving proposer membership (raw journal bytes)
+        vote_receipt: Vec<u8>,
         /// Proposer's vote nullifier (derived via vote_circuit)
         nullifier: [u8; 32],
     },
@@ -76,8 +76,8 @@ pub enum Instruction {
         create_key: [u8; 32],
         /// Member NPKs (needed by pre_tx hook for ZK proof)
         members: Vec<[u8; 32]>,
-        /// RISC0 receipt from vote_circuit proving voter membership
-        vote_receipt: Vec<u32>,
+        /// RISC0 receipt from vote_circuit proving voter membership (raw journal bytes)
+        vote_receipt: Vec<u8>,
         /// Voter's nullifier (prevents double-voting)
         nullifier: [u8; 32],
     },
@@ -90,8 +90,8 @@ pub enum Instruction {
         create_key: [u8; 32],
         /// Member NPKs (needed by pre_tx hook for ZK proof)
         members: Vec<[u8; 32]>,
-        /// RISC0 receipt from vote_circuit proving voter membership
-        vote_receipt: Vec<u32>,
+        /// RISC0 receipt from vote_circuit proving voter membership (raw journal bytes)
+        vote_receipt: Vec<u8>,
         /// Voter's nullifier (prevents double-voting)
         nullifier: [u8; 32],
     },
